requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
decouple==0.0.7
dj-rest-auth==7.0.0
Django==5.1.2
django-allauth==65.3.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
mypy-extensions==1.0.0
packaging==24.1
psycopg2==2.9.10
pydantic_core==2.23.4
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
types-PyYAML==6.0.12.20241221
typing_extensions==4.12.2
uritemplate==4.1.1
